Require v0.4.0 of stripes-core
I have some vague hope that this will fix FOLIO-520 -- if the demo
system is built on top of a git checkout of stripes-sample-platform
and otherwise relies on NPM packages, and if it automatically rebuilds
when a change is committed. At any rate, it's the best I can do until
Wayne and John wake up.